Refactor & Added TimeScheduler for Device State Updates. Added State to device api
