--- mosbach-tasks-rest-main/.vscode/launch.json
@@ -1,14 +1,0 @@
-{
-    "configurations": [
-        {
-            "type": "java",
-            "name": "Spring Boot-SmartHomeBackendApplication<smartHomeBackend>",
-            "request": "launch",
-            "cwd": "${workspaceFolder}",
-            "mainClass": "mosbach.dhbw.de.smarthome.SmartHomeBackendApplication",
-            "projectName": "smartHomeBackend",
-            "args": "",
-            "envFile": "${workspaceFolder}/.env"
-        }
-    ]
-}
