+++ mosbach-tasks-rest-main/src/main/java/mosbach/dhbw/de/smarthome/SmartHomeBackendApplication.java
@@ -3,14 +3,17 @@ package mosbach.dhbw.de.smarthome;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
+@EnableScheduling
 public class SmartHomeBackendApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(SmartHomeBackendApplication.class, args);
     }
 
+    //For rasberry pi use only
 //    @Bean
 //    public TomcatServletWebServerFactory servletContainer() {
 //        TomcatServletWebServerFactory tomcat = new TomcatServletWebServerFactory();
