+++ mosbach-tasks-rest-main/src/main/java/mosbach/dhbw/de/smarthome/controller/RoutineController.java
@@ -1,0 +1,68 @@
+package mosbach.dhbw.de.smarthome.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import mosbach.dhbw.de.smarthome.dto.ChangeRequest;
+import mosbach.dhbw.de.smarthome.dto.MessageToken;
+import mosbach.dhbw.de.smarthome.dto.RoutineDTO;
+
+@CrossOrigin(origins = "https://smarthomefrontend-surprised-oryx-bl.apps.01.cf.eu01.stackit.cloud", allowedHeaders = "*")
+@RestController
+@RequestMapping("/api/routine")
+public class RoutineController {
+    
+    @GetMapping(
+        consumes = {MediaType.APPLICATION_JSON_VALUE}
+    )
+    public ResponseEntity<?> getAllRoutines(@RequestBody MessageToken messageToken) { 
+        //TODO: Get All Routine Implementation
+        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+    }
+
+    @PostMapping(
+        consumes = {MediaType.APPLICATION_JSON_VALUE}
+    )
+    public ResponseEntity<?> createRoutine(@RequestBody RoutineDTO routinePostRequest) {
+        //TODO: Create Routine Implementation
+        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+    }
+
+    @GetMapping(
+        path = "/{id}",
+        consumes = {MediaType.APPLICATION_JSON_VALUE}
+    )
+    public ResponseEntity<?> getRoutine(@PathVariable String id, @RequestBody MessageToken tokenRequest) {
+        //TODO: Get Routine Implementation
+        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+    }
+
+    @DeleteMapping(
+        path = "/{id}",
+        consumes = {MediaType.APPLICATION_JSON_VALUE}
+    )
+    public ResponseEntity<?> deleteRoutine(@PathVariable String id, @RequestBody MessageToken tokenRequest) {
+        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+        //TODO: Delete Routine Implementation
+    }
+     
+    @PutMapping(
+        path = "/{id}",
+        consumes = {MediaType.APPLICATION_JSON_VALUE}
+    )
+    public ResponseEntity<?> changeRoutine(@PathVariable String id, @RequestBody ChangeRequest changeRequest) {
+        //TODO: Put Routine Implementation
+        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+    }
+
+}
