+++ mosbach-tasks-rest-main/src/main/java/mosbach/dhbw/de/smarthome/dto/smartthings/Data__1.java
@@ -8,17 +8,46 @@ import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-
+    "reason"
 })
 @Generated("jsonschema2pojo")
 public class Data__1 {
 
+    @JsonProperty("reason")
+    private String reason;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
+
+    /**
+     * No args constructor for use in serialization
+     * 
+     */
+    public Data__1() {
+    }
+
+    /**
+     * 
+     * @param reason
+     */
+    public Data__1(String reason) {
+        super();
+        this.reason = reason;
+    }
+
+    @JsonProperty("reason")
+    public String getReason() {
+        return reason;
+    }
+
+    @JsonProperty("reason")
+    public void setReason(String reason) {
+        this.reason = reason;
+    }
 
     @JsonAnyGetter
     public Map<String, Object> getAdditionalProperties() {
