+++ mosbach-tasks-rest-main/src/main/java/mosbach/dhbw/de/smarthome/dto/smartthings/GetFullStatusResponse.java
@@ -13,13 +13,19 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
+    "healthCheck",
+    "refresh",
     "switch"
 })
 @Generated("jsonschema2pojo")
-public class Switch {
+public class GetFullStatusResponse {
 
+    @JsonProperty("healthCheck")
+    private HealthCheck healthCheck;
+    @JsonProperty("refresh")
+    private Refresh refresh;
     @JsonProperty("switch")
-    private Switch__1 _switch;
+    private Switch _switch;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
 
@@ -27,25 +33,49 @@ public class Switch {
      * No args constructor for use in serialization
      * 
      */
-    public Switch() {
+    public GetFullStatusResponse() {
     }
 
     /**
      * 
+     * @param healthCheck
      * @param _switch
+     * @param refresh
      */
-    public Switch(Switch__1 _switch) {
+    public GetFullStatusResponse(HealthCheck healthCheck, Refresh refresh, Switch _switch) {
         super();
+        this.healthCheck = healthCheck;
+        this.refresh = refresh;
         this._switch = _switch;
     }
 
+    @JsonProperty("healthCheck")
+    public HealthCheck getHealthCheck() {
+        return healthCheck;
+    }
+
+    @JsonProperty("healthCheck")
+    public void setHealthCheck(HealthCheck healthCheck) {
+        this.healthCheck = healthCheck;
+    }
+
+    @JsonProperty("refresh")
+    public Refresh getRefresh() {
+        return refresh;
+    }
+
+    @JsonProperty("refresh")
+    public void setRefresh(Refresh refresh) {
+        this.refresh = refresh;
+    }
+
     @JsonProperty("switch")
-    public Switch__1 getSwitch() {
+    public Switch getSwitch() {
         return _switch;
     }
 
     @JsonProperty("switch")
-    public void setSwitch(Switch__1 _switch) {
+    public void setSwitch(Switch _switch) {
         this._switch = _switch;
     }
