+++ mosbach-tasks-rest-main/src/main/java/mosbach/dhbw/de/smarthome/model/Device.java
@@ -6,13 +6,15 @@ public class Device {
     private String type;
     private String status;
     private String location;
+    private String state;
 
     public Device(String id, String name, String type, String location) {
         this.id = id;
         this.name = name;
         this.type = type;
         this.location = location;
-        this.status = "Online";
+        this.status = "Offline";
+        this.state = "Off";
     }
 
     public String getName() {
@@ -51,5 +53,11 @@ public class Device {
         this.location = location;
     }
 
+    public String getState() {
+        return state;
+    }
 
+    public void setState(String state) {
+        this.state = state;
+    }
 }
