+++ mosbach-tasks-rest-main/src/main/java/mosbach/dhbw/de/smarthome/service/UserService.java
@@ -46,4 +46,8 @@ public class UserService {
             }
             return false;
         }
+
+        static List<User> getAllUsers() {
+            return users;
+        }
 }
